refactor loader scheduler components
